feat: added copy link component to notification, status and dashboard page
