+++ apps/tenant-management-webapp/src/app/pages/admin/dashboard/styled-components.ts
@@ -49,52 +49,3 @@ export const ListWrapper = styled.ul`
     margin-bottom: 0.5rem;
   }
 `;
-export const LinkCopyComponentWrapper = styled.div`
-  position: relative;
-  padding-top: 1.5rem;
-`;
-
-export const CopyLinkToolTipWrapper = styled.div`
-  .checkmark-icon {
-    display: inline-block;
-    margin-top: 0.5rem;
-    margin-left: 0.5rem;
-  }
-  .message {
-    font-size: 0.875rem;
-    display: inline;
-    position: absolute;
-    top: 0.3rem;
-    left: 2rem;
-  }
-  .URL-tooltip {
-    width: 30rem !important;
-    left: -5.5rem;
-    font-size: 10px;
-  }
-  p {
-    position: absolute;
-    background: var(--color-gray-100);
-    -webkit-border-radius: 10px;
-    -moz-border-radius: 10px;
-    border-radius: 30px;
-    width: 12rem;
-    height: 2.2rem;
-    top: -1.5rem;
-    left: 1rem;
-  }
-
-  p:before {
-    content: '';
-    position: absolute;
-    top: 2rem;
-    left: 6rem;
-    z-index: 1;
-    border: solid 15px transparent;
-    border-right-color: var(--color-gray-100);
-    border-top: 15px solid var(--color-gray-100);
-    border-right: 10px solid transparent;
-    border-left: 10px solid transparent;
-    border-bottom: none;
-  }
-`;
